requirements.txt
psycopg2-binary==2.9.9
requests==2.32.3
python-dotenv==1.2.2
dbt-core==1.9.0
dbt-postgres==1.9.0
